Redesigned with tailwind + tweaks
