--- src/components/Article/Article.css
@@ -1,34 +1,0 @@
-
-.article{
-    display: flex;
-    flex-direction: column;
-    place-content: center;
-    align-items:flex-start;
-    margin-top: 4rem;
-}
-.post-buttons{
-    align-self: flex-end;
-}
-.vote-buttons{
-    align-self: flex-end;
-}
-.comment-button{
-    align-self: flex-end;
-}
-.post-comment-button{
-    align-self: flex-end;
-}
-
-img{
-    width: 50%;
-    align-self: center;
-}
-.comment-container{
-    width: 100%;
-    border-radius: 20px;
-    opacity: 0.3;
-    background-color: grey;
-}
-textarea{
-    width: 95%;
-}
