--- src/components/ArticleCard/ArticleCard.css
@@ -1,39 +1,0 @@
-
-.article-card{
-    display: flex;
-    flex-direction:column ;
-    place-items: center;
-    max-width: 10rem;
-    min-height: 15rem;
-    max-height: 14rem;
-    
-}
-
-.article-card > h3 {
-    font-size: 0.8rem;
-    padding: 0.1rem;
-    margin: 0;
-    overflow:hidden;
-    text-wrap: wrap;
-    
-    
-}
-.article-card > h4 {
-    font-size: 0.7rem;
-    padding: 0.1rem;
-    margin: 0;
-    overflow: hidden;
-    text-wrap: wrap;
-}
-
-.article-card > img {
-    height: 6rem;
-    width: 6rem;
-    border-radius: 0.3rem;
-}
-
-.article-card > p {
-    font-size: 0.8rem;
-    padding: 0.1rem;
-    margin: 0;
-}
