--- src/components/Articles/Articles.css
@@ -1,34 +1,0 @@
-.articles-container {
-    margin-top: 3rem;
-    margin-bottom: 5rem;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-evenly;
-}
-.sort-filter-bar{
-    display: flex;
-    justify-content: space-evenly;
-    margin-top: 3.5rem;
-    border: solid 1px #ffffff;
-    opacity: 0.65;
-    border-radius: 20px;
-    width:100%;
-}
-.sort-filter-bar  select{
-    margin: 0.5rem;
-    background-color: #213547;
-    color: #ffffff;
-    
-}
-
-@media (prefers-color-scheme: light) {
-    .sort-filter-bar  select{
-        color: #213547;
-        background-color: #ffffff;
-        box-shadow: 0 0 3rem 3rem #ffffff;
-    }
-    .sort-filter-bar{
-        border: solid 1px #213547;
-    }
-
-}
