--- src/components/Carousel/Carousel.css
@@ -1,5 +1,0 @@
-.carousel-slide{
-    width:100%;
-    display: flex;
-    flex-direction: column;
-}
