--- src/components/CommentCard/CommentCard.css
@@ -1,24 +1,0 @@
-.comment{
-    margin: 20px;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    background-color: grey;
-    border-radius: 20px;
-    padding: 10px;
-    opacity: 80%;
-}
-.comment>div{
-    display: flex;
-    width: 100%;
-    justify-content: space-between;
-}
-
-.comment-author{
-    align-self: flex-end;
-    order: 2;
-   
-}
-.comment-buttons{
-align-self: flex-start;
-}
