--- src/components/Footer/Footer.css
@@ -1,4 +1,0 @@
-footer{
-    position: relative;
-    bottom: 0;
-}
