--- src/components/Footer/Footer.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Footer = () => {
-  return (
-    <footer>Footer</footer>
-  )
-}
