+++ src/components/Header.jsx
@@ -1,0 +1,9 @@
+/* eslint-disable react/prop-types */
+
+export const Header = ({ Navbar }) => {
+  return (
+
+    <header className='  z-10 w-full flex-row place-content-center bg-bkg text-content rounded-lg   fixed shadow-2xl'>{Navbar}</header>
+
+  )
+}
