--- src/components/Header/Header.css
@@ -1,34 +1,0 @@
-* {
-    box-sizing: border-box;
-}
-
-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    padding: 0.5rem;
-    height: 5rem;
-    font-size: 2rem;
-    border-bottom: 1px solid white;
-    z-index: 1;
-    background-color: #242424;
-    box-shadow: 0 0 3rem 3rem #242424;
-}
-
-header > img {
-    height: 4rem;
-    width: 4rem;
-}
-
-@media (prefers-color-scheme: light) {
-    header {
-        color: #213547;
-        background-color: #ffffff;
-        box-shadow: 0 0 3rem 3rem #ffffff;
-    }
-
-}
