--- src/components/Header/Header.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import './Header.css'
-
-export const Header = ({Navbar}) => {
-  return (
-
-    <>{Navbar}</>
-
-  )
-}
