--- src/components/LoadingSpinner/LoadingSpinner.jsx
@@ -1,19 +1,0 @@
-import React from 'react'
-
-export const LoadingSpinner = () => {
-  return (
-    <div className="sk-circle">
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-    <div className="sk-circle-dot"></div>
-</div>  )
-}
