--- src/components/Login/Login.css
@@ -1,24 +1,0 @@
-input,
-label,
-select,
-button,
-textarea
-{
-	margin:0;
-	border:0;
-	padding:0;
-	display:inline-block;
-	vertical-align:middle;
-	white-space:normal;
-	background:none;
-	line-height:1;
-	
-	/* Browsers have different default form fonts */
-	font-size:13px;
-	font-family:Arial;
-}
-.login-section{
-
-position: absolute;
-border-bottom: none !important;
-}
