--- src/components/Navbar/Navbar.css
@@ -1,100 +1,0 @@
-nav {
-    display: flex;
-    justify-content: center;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    border-top: 1px solid white;
-    padding: 0.2rem;
-    z-index: 1;
-    background-color: #242424;
-    box-shadow: 0 0 2rem 1.5rem #242424;
-
-    border: none;
-}
-
-nav a>button {
-    width: 25vw;
-    min-width: 150px;
-}
-
-nav a {
-    display: flex;
-    place-content: center;
-    width: 20vw;
-    max-width: 200px;
-    font-size: 1rem;
-    margin: 0.7rem 1.2rem;
-}
-
-.navBtns {
-    display: flex;
-}
-
-.burgerBtn {
-    display: none;
-}
-
-.userBtn {
-    display: flex;
-    position: fixed;
-    right: 0;
-    top: 0;
-    width: 55px;
-    padding: 0.2rem;
-    margin: 0.5rem;
-
-}
-
-
-
-@media (prefers-color-scheme: light) {
-    nav {
-        color: #213547;
-        background-color: #ffffff;
-        box-shadow: 0 0 3rem 3rem #ffffff;
-        width: 100vw;
-    }
-
-}
-
-@media screen and (max-width: 830px) {
-    .navBtns {
-        position: absolute;
-        display: flex;
-        left: 0;
-        top: 71px;
-        width: 60vw;
-        flex-direction: column;
-        align-items: flex-start;
-    }
-
-    .navBtns a {
-        width: fit-content;
-        margin: 0.1rem;
-    }
-
-    .false {
-        display: none;
-    }
-
-    .burgerBtn {
-        display: flex;
-        position: fixed;
-        width: 55px;
-        padding: 0.2rem;
-        margin: 0.5rem;
-        top: 1px;
-
-    }
-
-    .userBtn {
-        top: 1px;
-        left: auto;
-    }
-
-    nav {
-        justify-content: space-between;
-    }
-}
