--- src/components/UserCard/UserCard.css
@@ -1,18 +1,0 @@
-.user-card-container {
-    display: flex;
-    margin: 1rem;
-    background-color: grey;
-    border-radius: 20px;
-    max-height: 5rem;
-    width: 20rem;
-
-}
-
-.user-img {
-    position: relative;
-    max-width: 8rem;
-    max-height: inherit;
-    object-fit:contain;
-    overflow: hidden;
-   
-}
