+++ src/main.jsx
@@ -6,10 +6,8 @@ import { BrowserRouter } from 'react-router-dom'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-
     <BrowserRouter>
     <App />
     </BrowserRouter>
-
   </React.StrictMode>,
 )
